+++ rest_framework/serializers/__init__.py
@@ -4,7 +4,7 @@ Serializers.
 """
 from .serializers import Serializer, ListSerializer
 from .fields import (
-    BooleanField, CharField, IntegerField, FloatField, ListField,
+    BooleanField, BooleanNullField, CharField, IntegerField, FloatField, ListField,
     TimeField, DateField, DateTimeField,
     JsonField, DictField,
     SerializerMethodField,
@@ -14,14 +14,14 @@ from .exceptions import ValidationError
 
 __all__ = (
     # fields
-    'BooleanField', 'CharField', 'IntegerField', 'FloatField', 'ListField',
-    'TimeField', 'DateField', 'DateTimeField',
-    'JsonField', 'DictField',
-    'SerializerMethodField',
+    BooleanField, BooleanNullField, CharField, IntegerField, FloatField, ListField,
+    TimeField, DateField, DateTimeField,
+    JsonField, DictField,
+    SerializerMethodField,
 
     # serializers
-    'Serializer', 'ListSerializer',
+    Serializer, ListSerializer,
 
     # exceptions
-    'ValidationError',
+    ValidationError,
 )